# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gestão Escolar Api",
        "description": "A gestão escolar com API envolve a integração de sistemas e aplicativos para otimizar a administração escolar, permitindo a troca de dados e informações de forma eficiente entre diferentes componentes do ambiente educacional, como registros acadêmicos, matrículas, sistemas de pagamento e comunicação com pais e alunos.",
        "contact": {
            "email": "adilson2012jose@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.4.0"
    },
    "paths": {
...
        "/api/Estudante/dadosdoestudantelogado": {
            "get": {
                "tags": [
                    "Estudante"
                ],
                "summary": "Detalhes do Estudante da Classe Actual",
                "operationId": "23701efe881a68808b6ba4ed986247bd",
                "responses": {
                    "200": {
                        "description": "Dados Detalhados de Current Anolectivo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "telefoneAlternativo": {
                                                "type": "string",
                                                "example": "926551976"
                                            },
                                            "numeroDotelefone": {
                                                "type": "string",
                                                "example": "915882240"
                                            },
                                            "nomeMae": {
                                                "type": "string",
                                                "example": "Makiande Jose Kiala"
                                            },
                                            "nomePai": {
                                                "type": "string",
                                                "example": "Nzilau Fernando Miguel"
                                            },
                                            "municipio_id": {
                                                "type": "string",
                                                "example": "Maianga"
                                            },
                                            "pais": {
                                                "type": "string",
                                                "example": "Angola"
                                            },
                                            "provincia_id": {
                                                "type": "string",
                                                "example": "Luanda"
                                            },
                                            "genero_id": {
                                                "type": "string",
                                                "example": "Masculino"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Estudante/disciplinas/anolectivo/{ano}": {
            "get": {
                "tags": [
                    "Estudante"
                ],
                "summary": "lista de Disciplina Determinado Ano Lectivo",
                "operationId": "49f3e60cb2e5ede59a99f92c64aaaece",
                "parameters": [
                    {
                        "name": "ano",
                        "in": "path",
                        "description": "ID do ano letivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Disciplinas de um determiando Ano Lectivo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "Professor Professor": {
                                                "type": "string",
                                                "example": "Joao Miguel"
                                            },
                                            "nomeCurso": {
                                                "type": "string",
                                                "example": "Ciências Físicas e Biológica"
                                            },
                                            "nomeDisciplina": {
                                                "type": "string",
                                                "example": "INGLÊS"
                                            },
                                            "nomePeriodo": {
                                                "type": "string",
                                                "example": "Tarde"
                                            },
                                            "nomeSala": {
                                                "type": "string",
                                                "example": "Sala 9"
                                            },
                                            "nomeTurma": {
                                                "type": "string",
                                                "example": "B"
                                            },
                                            "classe_name": {
                                                "type": "string",
                                                "example": "10 Classe"
                                            },
                                            "ano_lectivo": {
                                                "type": "string",
                                                "example": "2023\\/2024"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Estudante/disciplinas/gradecurricular": {
            "get": {
                "tags": [
                    "Estudante"
                ],
                "summary": "lista de Grade curricular Determinado",
                "operationId": "64968d19670658115fea7b72fe438397",
                "responses": {
                    "200": {
                        "description": "Lista de Grade curricular",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "Professor Professor": {
                                                "type": "string",
                                                "example": "Joao Miguel"
                                            },
                                            "nomeCurso": {
                                                "type": "string",
                                                "example": "Ciências Físicas e Biológica"
                                            },
                                            "nomeDisciplina": {
                                                "type": "string",
                                                "example": "INGLÊS"
                                            },
                                            "nomePeriodo": {
                                                "type": "string",
                                                "example": "Tarde"
                                            },
                                            "nomeSala": {
                                                "type": "string",
                                                "example": "Sala 9"
                                            },
                                            "nomeTurma": {
                                                "type": "string",
                                                "example": "B"
                                            },
                                            "classe_name": {
                                                "type": "string",
                                                "example": "10 Classe"
                                            },
                                            "ano_lectivo": {
                                                "type": "string",
                                                "example": "2023\\2024"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Estudante/disciplinas/pendentes/anolectivo/{ano}": {
            "get": {
                "tags": [
                    "Estudante"
                ],
                "summary": "lista de Disciplinas Pendentes or para Recurso",
                "operationId": "69dd77bb05ada699a655c23a47cf57ad",
                "parameters": [
                    {
                        "name": "ano",
                        "in": "path",
                        "description": "ID do ano letivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "lista de Disciplinas Pendentes or para Recurso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "Professor Professor": {
                                                "type": "string",
                                                "example": "Joao Miguel"
                                            },
                                            "nomeCurso": {
                                                "type": "string",
                                                "example": "Ciências Físicas e Biológica"
                                            },
                                            "nomeDisciplina": {
                                                "type": "string",
                                                "example": "INGLÊS"
                                            },
                                            "nomePeriodo": {
                                                "type": "string",
                                                "example": "Tarde"
                                            },
                                            "nomeSala": {
                                                "type": "string",
                                                "example": "Sala 9"
                                            },
                                            "nomeTurma": {
                                                "type": "string",
                                                "example": "B"
                                            },
                                            "classe_name": {
                                                "type": "string",
                                                "example": "10 Classe"
                                            },
                                            "ano_lectivo": {
                                                "type": "string",
                                                "example": "2023\\/2024"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}